fix(review): address Copilot PR feedback
- Also intercept '-v' in the early --version handler (consistent with '-V' and '--version')
- Ignore empty/whitespace SCFZ_UPDATE_CACHE_DIR and fall back to homedir()
- Move --version update notification E2E tests to scfz-version.test.ts so they don't run in the Windows/macOS smoke job (test:e2e:smoke:windows targets scfz-general.test.ts only)

Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
